add helper function for building redirect URLs
