+++ app/scripts/utils/format.js
@@ -2,6 +2,7 @@
 import React from 'react';
 import moment from 'moment';
 import numeral from 'numeral';
+import url from 'url';
 import { Link } from 'react-router';
 
 export const nullValue = '--';
@@ -164,4 +165,20 @@ export const deleteText = function (name) {
 
 export const rerunText = function (name) {
   return `Are you sure you want to rerun ${name}?`;
+};
+
+export const buildRedirectUrl = function ({ origin, pathname, hash }) {
+  const hasQuery = hash.indexOf('?');
+  if (hasQuery !== -1) {
+    hash = hash.substr(hash.indexOf('#') + 1);
+    const parsedUrl = url.parse(hash, true);
+    // Remove any ?token query parameter to avoid polluting
+    // the login link
+    delete parsedUrl.query.token;
+    hash = `#${url.format({
+      pathname: parsedUrl.pathname,
+      query: parsedUrl.query
+    })}`;
+  }
+  return encodeURIComponent(url.resolve(origin, pathname) + hash);
 };
